GPU Test Stage
Still need to find a version of blasGemm that can be used on the GPU for
easier testing, so then we can reuse the testing code from homework 1.
Otherwise, we might need to implement our own version, which will be
much slower but should still be correct.

+++ GPU/gpu.cu
@@ -89,10 +89,10 @@ void init_spgemm(const sparse_mat_t &A, const sparse_mat_t &B)
     cudaMalloc(&d_B_col_indices, sizeof(int) * (B.values.size()));
     cudaMemcpy(d_B_col_indices, B.col_indices.data(), sizeof(int) * (B.values.size()), cudaMemcpyHostToDevice);
 
-    cudaMalloc(&d_A_values, sizeof(int) * (A.values.size()));
-    cudaMemcpy(d_A_values, A.values.data(), sizeof(int) * (A.values.size()), cudaMemcpyHostToDevice);
-    cudaMalloc(&d_B_values, sizeof(int) * (B.values.size()));
-    cudaMemcpy(d_B_values, B.values.data(), sizeof(int) * (B.values.size()), cudaMemcpyHostToDevice);
+    cudaMalloc(&d_A_values, sizeof(double) * (A.values.size()));
+    cudaMemcpy(d_A_values, A.values.data(), sizeof(double) * (A.values.size()), cudaMemcpyHostToDevice);
+    cudaMalloc(&d_B_values, sizeof(double) * (B.values.size()));
+    cudaMemcpy(d_B_values, B.values.data(), sizeof(double) * (B.values.size()), cudaMemcpyHostToDevice);
 
     printf("The number of blocks %d \n", blks);
 }
